+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - chapter-9.rb [examples from chapter-9]
 - one-billion-seconds.rb [todo from chapter-9]
 - happy-birthday.rb [todo from chapter-9]
+- orange-tree.rb [todo from chapter-9]
 
 **Updated :**